+++ CHANGELOG.md
@@ -14,6 +14,13 @@ This is the official BubbleOS changelog! All features will be recorded in _'grou
 ### Changed/Fixed Features
 
 - Improved the selection algorithm of the `tips` command, so that it is impossible to fail finding a tip, and when there are no more tips to show, it will reset automatically.
+- The color for memory in `sysinfo` also has yellow text for if the memory usage is around the middle.
+- The information in the `sysinfo` command now shows bolded instead of italicized.
+- Increased the timeout time in the `ping` command from 5 seconds to 15 seconds.
+- Fixed an issue where the `ping` command would crash BubbleOS if the status code returned was not 200, and also if the connection to the address was refused.
+- Fixed an issue where using the `dirtree` command would crash BubbleOS.
+- Fixed an issue where the `bub` command would not show the command being executed if it was the `exit` command or the `bub` command running the same file.
+- Fixed an issue where changing the directory in a `.bub` file would keep its changes after it was done executing.
 - Fixed a consistency issue where the user information had its values bolded instead of italicized in the `sysinfo` command.
 
 ### Removed Features